requirements.txt
asgiref==3.8.1
asttokens==2.2.1
Babel==2.11.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.3
cx-Oracle==8.3.0
debugpy==1.6.7
decorator==4.4.2
distlib==0.3.8
Django==5.0.6
django-appconf==1.0.6
django-imagekit==5.0.0
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.14.0
future==1.0.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
moviepy==1.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
nest-asyncio==1.5.7
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pilkit==3.0
pillow==10.3.0
platformdirs==3.10.0
proglog==0.1.10
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==306
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
tkcalendar==1.6.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
